+++ app/tests/v1/test_auth.py
@@ -2,7 +2,7 @@ import pytest
 from flask import json
 from app import create_app
 from app.api.v1.views.auth import Users
-from app.tests import sign_in_admin_helper, sign_in_helper, sign_up_helper
+from app.tests.test_helpers import sign_in_admin_helper, sign_in_helper, sign_up_helper
 
 testusers = Users()
