+++ app/tests/v1/test_sales.py
@@ -3,8 +3,8 @@ from flask import json
 from app import create_app
 from app.api.v1.views.items import Items
 from app.api.v1.views.sales import Sales
-from app.api.v1.views.auth import Users
-from app.tests import make_sale_helper, sign_in_admin_helper, sign_in_helper, add_items_helper, yield_test_client, sign_up_helper
+from app.api.v1.views.auth import Users, users
+from app.tests.test_helpers import make_sale_helper, sign_in_admin_helper, sign_in_helper, sign_in_helper_2, add_items_helper, yield_test_client, sign_up_sign_in_user_helper
 
 testitems = Items()
 testsales = Sales()
@@ -100,17 +100,15 @@ sample_sale=[{
             "sale_items": [
     	                    {
     		                "item_id":"1",
-    		                "quantity":"4"
+    		                "quantity":"1"
     	                    },
     	                    {
     		                "item_id":"2",
-    		                "quantity":"9"
+    		                "quantity":"1"
     	                    }
                            ]
             }
         ]
-
-
 
 
 '''-------------------------------------------------------------------------------------------------------------------------------'''
@@ -127,7 +125,6 @@ def test_sales_retrive_all_no_sale():
 
 def test_sales_retrive_all_successfully():
     test_client = app.test_client()
-    sign_up_helper(test_client)
     make_sale_helper(test_client)
     sign_in_admin_helper(test_client)
     response = test_client.get('/api/v1/sales',content_type='application/json')
@@ -140,56 +137,56 @@ def test_sales_retrive_all_successfully():
 
 def test_sales_quantity_not_digit():
     test_client = app.test_client()
-    sign_in_helper(test_client)
+    sign_in_helper_2(test_client)
     response = test_client.post('/api/v1/make_sale', data=json.dumps(sample_sale[0]) ,content_type='application/json')
     assert(response.status_code==400)
     
 
 def test_sales_item_id_not_digit():
     test_client = app.test_client()
-    sign_in_helper(test_client)
+    sign_in_helper_2(test_client)
     response = test_client.post('/api/v1/make_sale', data=json.dumps(sample_sale[1]) ,content_type='application/json')
     assert(response.status_code==400)
     
 
 def test_sales_item_id_empty():
     test_client = app.test_client()
-    sign_in_helper(test_client)
+    sign_in_helper_2(test_client)
     response = test_client.post('/api/v1/make_sale', data=json.dumps(sample_sale[2]) ,content_type='application/json')
     assert(response.status_code==400)
     
 
 def test_sales_quantity_empty():
     test_client = app.test_client()
-    sign_in_helper(test_client)
+    sign_in_helper_2(test_client)
     response = test_client.post('/api/v1/make_sale', data=json.dumps(sample_sale[3]) ,content_type='application/json')
     assert(response.status_code==400)
 
 
 def test_sales_quantity_more_than_available_stock():
     test_client = app.test_client()
-    sign_in_helper(test_client)
+    sign_in_helper_2(test_client)
     response = test_client.post('/api/v1/make_sale', data=json.dumps(sample_sale[4]) ,content_type='application/json')
-    assert(response.status_code==400)
+    assert(response.status_code==406)
 
 
 def test_sales_payment_method_empty():
     test_client = app.test_client()
-    sign_in_helper(test_client)
+    sign_in_helper_2(test_client)
     response = test_client.post('/api/v1/make_sale', data=json.dumps(sample_sale[5]) ,content_type='application/json')
     assert(response.status_code==406)
     
 
 def test_sales_sale_items_empty():
     test_client = app.test_client()
-    sign_in_helper(test_client)
+    sign_in_helper_2(test_client)
     response = test_client.post('/api/v1/make_sale', data=json.dumps(sample_sale[6]) ,content_type='application/json')
     assert(response.status_code==406)
     
 
 def test_place_sale_successfully():
     test_client = app.test_client()
-    sign_in_helper(test_client)
+    sign_in_helper_2(test_client)
     response = test_client.post('/api/v1/make_sale', data=json.dumps(sample_sale[7]) ,content_type='application/json')
     json.loads(response.data)
     assert(response.status_code==201)
@@ -215,7 +212,7 @@ def test_get_sale_not_created():
 
 def test_get_sale_successfully():
     test_client = app.test_client()
-    make_sale_helper(test_client)
+    #make_sale_helper(test_client)
     sign_in_admin_helper(test_client)
     response = test_client.get('/api/v1/sales/1' ,content_type='application/json')
     assert(response.status_code == 200)
